cr/bulk-import/src/main/java/org/lilyproject/repository/bulk/BulkIngester.java
/*
 * Copyright 2012 NGDATA nv
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.lilyproject.repository.bulk;

import java.io.Closeable;
import java.io.IOException;
import java.io.OutputStream;
import java.util.List;
import java.util.Set;

import com.google.common.collect.Sets;
import org.apache.hadoop.conf.Configuration;
import org.apache.hadoop.hbase.HBaseConfiguration;
import org.apache.hadoop.hbase.client.HTableInterface;
import org.apache.hadoop.hbase.client.Put;
import org.lilyproject.repository.api.Blob;
import org.lilyproject.repository.api.BlobAccess;
import org.lilyproject.repository.api.BlobException;
import org.lilyproject.repository.api.BlobManager;
import org.lilyproject.repository.api.BlobReference;
import org.lilyproject.repository.api.BlobStoreAccess;
import org.lilyproject.repository.api.FieldType;
import org.lilyproject.repository.api.FieldTypes;
import org.lilyproject.repository.api.IdGenerator;
import org.lilyproject.repository.api.QName;
import org.lilyproject.repository.api.Record;
import org.lilyproject.repository.api.RecordFactory;
import org.lilyproject.repository.api.RecordId;
import org.lilyproject.repository.api.Repository;
import org.lilyproject.repository.api.RepositoryException;
import org.lilyproject.repository.api.RepositoryManager;
import org.lilyproject.repository.api.TypeManager;
import org.lilyproject.repository.impl.HBaseRepository;
import org.lilyproject.repository.impl.HBaseRepositoryManager;
import org.lilyproject.repository.impl.HBaseTypeManager;
import org.lilyproject.repository.impl.RecordFactoryImpl;
import org.lilyproject.repository.impl.id.IdGeneratorImpl;
import org.lilyproject.util.exception.ExceptionUtil;
import org.lilyproject.util.hbase.HBaseTableFactory;
import org.lilyproject.util.hbase.HBaseTableFactoryImpl;
import org.lilyproject.util.hbase.LilyHBaseSchema;
import org.lilyproject.util.repo.RecordEvent;
import org.lilyproject.util.repo.RecordEvent.Type;
import org.lilyproject.util.zookeeper.ZkUtil;
import org.lilyproject.util.zookeeper.ZooKeeperItf;
import org.python.google.common.collect.Lists;

/**
 * Writes Lily records in bulk to HBase. Also provides methods for creating HBase {@code Put} or
 * {@code KeyValue} object for alternative methods of writing (e.g. via MapReduce).
 */
public class BulkIngester implements Closeable {
    
    public static final int PUT_BUFFER_SIZE = 1000;

    private HBaseRepository hbaseRepo;
    private RepositoryManager repositoryManager;
    private RecordFactory recordFactory;
    private HTableInterface recordTable;
    private FieldTypes fieldTypes;
    private List<Put> putBuffer = Lists.newArrayListWithCapacity(PUT_BUFFER_SIZE);

    /**
     * Factory method for creation of a {@code BulkIngester} that operates on the default repository table.
     * 
     * @param zkConnString Connection string for ZooKeeper
     * @param timeout ZooKeeper session timeout
     * @return a new BulkIngester
     */
    public static BulkIngester newBulkIngester(String zkConnString, int timeout) {
        return newBulkIngester(zkConnString, timeout, null);
    }
    
    /**
     * Factory method for creation of a {@code BulkIngester that operates on a non-default repository table.
     * 
     * @param zkConnString connection string for ZooKeeper
     * @param timeout ZooKeeper session timeout
     * @param tableName name of the repository table to write to
     */
    public static BulkIngester newBulkIngester(String zkConnString, int timeout, String tableName) {
        try {
            ZooKeeperItf zk = ZkUtil.connect(zkConnString, timeout);
            Configuration conf = HBaseConfiguration.create();
            conf.set("hbase.zookeeper.quorum", zkConnString);
            HBaseTableFactory hbaseTableFactory = new HBaseTableFactoryImpl(conf);
            IdGenerator idGenerator = new IdGeneratorImpl();
            TypeManager typeManager = new HBaseTypeManager(idGenerator, conf, zk, hbaseTableFactory);
            RecordFactory recordFactory = new RecordFactoryImpl(typeManager, idGenerator);
            
            @SuppressWarnings("resource") // RepositoryManager gets closed in BulkIngester.close
            RepositoryManager repositoryManager = new HBaseRepositoryManager(typeManager, idGenerator,
                        recordFactory, hbaseTableFactory, new BlobsNotSupportedBlobManager());

            // FIXME Blobs aren't really supported here (no BlobManager is created), but null is
            // just passed in as the BlobManager so we'll probably get some mystery NPEs if Blobs
            // are used. We should either support blobs, or at least forcefully disallow them
            HBaseRepository hbaseRepository;
            if (tableName != null) {
                hbaseRepository = (HBaseRepository)repositoryManager.getRepository(tableName);
            } else {
                hbaseRepository = (HBaseRepository)repositoryManager.getDefaultRepository();
            }
            return new BulkIngester(hbaseRepository, LilyHBaseSchema.getRecordTable(hbaseTableFactory, hbaseRepository.getTableName()),
                    typeManager.getFieldTypesSnapshot());
        } catch (Exception e) {
            ExceptionUtil.handleInterrupt(e);
            throw new RuntimeException(e);
        }
    }

    BulkIngester(HBaseRepository hbaseRepo, HTableInterface recordTable, FieldTypes fieldTypes) {
        this.hbaseRepo = hbaseRepo;
        this.repositoryManager = hbaseRepo.getRepositoryManager();
        this.recordFactory = repositoryManager.getRecordFactory();
        this.recordTable = recordTable;
        this.fieldTypes = fieldTypes;
    }

    /**
     * Factory method for creation of Records, with the same semantics as
     * {@link Repository#newRecord()}.
     * 
     * @return A newly-created record
     */
    public Record newRecord() {
        return recordFactory.newRecord();
    }

    /**
     * Same as {@link Repository#getIdGenerator()}.
     * 
     * @return The IdGenerator of the underlying repository
     */
    public IdGenerator getIdGenerator() {
        return hbaseRepo.getIdGenerator();
    }

    /**
     * Write a single record directly to Lily, circumventing any indexing or other secondary actions
     * that are performed when using the standard Lily API.
     * <p>
     * <b>WARNING:</b>This method is not thread-safe.
     * <p>
     * Puts are first written to a buffer, which is flushed when it reaches {@link BulkIngester#PUT_BUFFER_SIZE}.
     * 
     * @param record Record to be written
     * @throws IOException
     */
    public void write(Record record) throws InterruptedException, RepositoryException, IOException {
        putBuffer.add(buildPut(record));
        if (putBuffer.size() == PUT_BUFFER_SIZE) {
            flush();
        }
    }
    
    /**
     * Flush buffered Puts to the Lily record table.
     * <p>
     * This method is not thread-safe.
     * 
     * @throws IOException
     */
    public void flush() throws IOException{
        if (!putBuffer.isEmpty()) {
            recordTable.put(Lists.newArrayList(putBuffer));
            putBuffer.clear();
        }
    }

    /**
     * Build the {@code Put} that represents a record for inserting into HBase.
     * @param record The record to be translated into an HBase {@code Put}
     * @return Put which can be directly written to HBase
     */
    public Put buildPut(Record record) throws InterruptedException, RepositoryException {
        RecordEvent recordEvent = new RecordEvent();
        recordEvent.setType(Type.CREATE);
        if (record.getId() == null) {
            record.setId(getIdGenerator().newRecordId());
        }
        return hbaseRepo.buildPut(record, 1L, fieldTypes, recordEvent, Sets.<BlobReference> newHashSet(),
                Sets.<BlobReference> newHashSet(), 1L);
    }
    
    @Override
    public void close() throws IOException {
        flush();
        repositoryManager.close();
    }
    
    
    /**
     * BlobManager that ensures that blobs aren't supported in bulk import.
     * <p>
     * In the future, it may be interesting to support blobs, but at the moment it seems as this
     * would likely lead to poor import performance.
     */
    private static class BlobsNotSupportedBlobManager implements BlobManager {

        private static final String NOT_SUPPORTED_MESSAGE = "Blobs are not supported for bulk imports";
        
        @Override
        public void incubateBlob(byte[] blobKey) throws IOException {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }

        @Override
        public Set<BlobReference> reserveBlobs(Set<BlobReference> blobs) throws IOException {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }

        @Override
        public void handleBlobReferences(RecordId recordId, Set<BlobReference> referencedBlobs,
                Set<BlobReference> unReferencedBlobs) {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }

        @Override
        public OutputStream getOutputStream(Blob blob) throws BlobException {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }

        @Override
        public BlobAccess getBlobAccess(Record record, QName fieldName, FieldType fieldType, int... indexes)
                throws BlobException {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }

        @Override
        public void register(BlobStoreAccess blobStoreAccess) {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }

        @Override
        public void delete(byte[] blobKey) throws BlobException {
            throw new UnsupportedOperationException(NOT_SUPPORTED_MESSAGE);
        }
        
        
    }

}
